make 8a, 8c, 8l build in go environment. ,s/int32/long/g in 8a, 8c, 8l. delete dead code. move enam.c, 8.out.h, mkenam from 8c to 8l.
R=r
DELTA=1850  (581 added, 983 deleted, 286 changed)
OCL=22119
CL=22129

+++ src/cmd/8c/cgen.c
@@ -39,7 +39,7 @@ cgen(Node *n, Node *nn)
 	Prog *p1;
 	Node nod, nod1, nod2, nod3, nod4;
 	int o, hardleft;
-	long v, curs;
+	int32 v, curs;
 	vlong c;
 
 	if(debug['g']) {
@@ -1031,7 +1031,7 @@ cgen(Node *n, Node *nn)
 			diag(n, "DOT and no offset");
 			break;
 		}
-		nod.xoffset += (long)r->vconst;
+		nod.xoffset += (int32)r->vconst;
 		nod.type = n->type;
 		cgen(&nod, nn);
 		break;
@@ -1133,7 +1133,7 @@ void
 reglcgen(Node *t, Node *n, Node *nn)
 {
 	Node *r;
-	long v;
+	int32 v;
 
 	regialloc(t, n, nn);
 	if(n->op == OIND) {
@@ -1217,7 +1217,7 @@ boolgen(Node *n, int true, Node *nn)
 	int o;
 	Prog *p1, *p2;
 	Node *l, *r, nod, nod1;
-	long curs;
+	int32 curs;
 
 	if(debug['g']) {
 		prtree(nn, "boolgen lhs");
@@ -1429,7 +1429,7 @@ boolgen(Node *n, int true, Node *nn)
 }
 
 void
-sugen(Node *n, Node *nn, long w)
+sugen(Node *n, Node *nn, int32 w)
 {
 	Prog *p1;
 	Node nod0, nod1, nod2, nod3, nod4, *h, *l, *r;
@@ -1500,7 +1500,7 @@ sugen(Node *n, Node *nn, long w)
 			diag(n, "DOT and no offset");
 			break;
 		}
-		nod1.xoffset += (long)r->vconst;
+		nod1.xoffset += (int32)r->vconst;
 		nod1.type = n->type;
 		sugen(&nod1, nn, w);
 		break;
